feat: refactor deleteOrders logic to handle multiple deletions and improve error handling
- Updated deleteOrders controller to accept an id and deleteMany flag.
- Enhanced deleteOrder service to support deleting multiple completed orders.
- Improved response messages for successful deletions and error handling.

+++ src/services/orders.service.ts
@@ -64,16 +64,23 @@ export const updateOrder = async (id: string, status: string) => {
     }
 }
 
-export const deleteOrder = async (id: string) => {
+export const deleteOrder = async (id: string, deleteMany: boolean = false) => {
     try {
-        
-        return await prisma.order.deleteMany({
-            where:{
-                status:"Concluído"
-            }
-        })
+        if (deleteMany) {
+            return await prisma.order.deleteMany({
+                where: {
+                    status: "Concluído"
+                }
+            });
+        }
 
+        return await prisma.order.delete({
+            where: {
+                id: id
+            }
+        });
     } catch (error) {
-        console.error(error)
+        console.error(error);
+        throw error;
     }
 }
